% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerplot-package.R
\name{DFA_plot}
\alias{DFA_plot}
\title{A function to plot predicted groups assignments.}
\usage{
DFA_plot(data)
}
\arguments{
\item{data}{predicted group assignments derived from the DFA_pred function.}
}
\description{
This function takes prediction data and plots it using ggplot.
}
\keyword{Prediction}
\keyword{plot}
