% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerplot-package.R
\name{Partition}
\alias{Partition}
\title{A function to partition data sets for use in LDFA.}
\usage{
Partition(x, p = 0.8)
}
\arguments{
\item{x}{your data}

\item{p}{fraction of observations allocated to training set}
}
\description{
These functions allow you to partition the observations in your data set in to test and training subsets.
}
\keyword{partition}
